Fix profile card story
