--- src/profile-card/profile-card.css
@@ -1,69 +1,0 @@
-:host {
-  display: inline-block;
-  --width: 215px;
-}
-.profile-card {
-  display: flex;
-  flex-direction: column;
-  width: var(--width);
-
-  border-radius: 5px;
-  overflow: hidden;
-}
-figure {
-  border-radius: 0.5rem;
-  height: var(--width);
-  margin: 0 0 0.5rem 0;
-  padding: 0;
-}
-
-img {
-  background-color: #ffffff;
-  filter: grayscale(1);
-  width: 100%;
-  object-fit: cover;
-  aspect-ratio: 1 / 1;
-  animation: fadeIn 1s cubic-bezier(0.46, 0.03, 0.05, 0.99);
-}
-
-@keyframes fadeIn {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-
-.profile-card__info {
-  display: flex;
-  background-color: var(--miles_secondary_four);
-  padding: 1.2rem;
-  min-height: 140px;
-  flex-direction: column;
-}
-
-.jobtitle {
-  flex-grow: 1;
-}
-
-.jobtitle p {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  display: -webkit-box;
-  -webkit-line-clamp: 2; /* number of lines to show */
-  line-clamp: 2;
-  -webkit-box-orient: vertical;
-}
-
-.description {
-  display: flex;
-  flex-direction: row;
-  align-items: baseline;
-  gap: 0.2rem;
-}
-h3,
-p {
-  margin: 0;
-  color: var(--miles_primary_light);
-}
