+++ src/profile-card/profile-card.scss
@@ -1,0 +1,79 @@
+:host {
+  display: inline-block;
+}
+
+*, *::before, *::after {
+  box-sizing: border-box;
+}
+
+.profile-card {
+  border-radius: 30px;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+  width: 100%;
+
+  @media (min-width: 600px) {
+    width: 260px;
+  }
+
+  &__info {
+    display: flex;
+    background-color: var(--miles_secondary_four);
+    padding: 1.2rem;
+    min-height: 140px;
+    flex-direction: column;
+  }
+
+  figure {
+    border-radius: 0.5rem;
+    aspect-ratio: 1/1;
+    margin: 0 0 0.25rem 0;
+    padding: 0;
+
+    img {
+      animation: fadeIn 1s cubic-bezier(0.46, 0.03, 0.05, 0.99);
+      aspect-ratio: 1 / 1;
+      display: block;
+      background-color: #ffffff;
+      filter: grayscale(1);
+      object-fit: cover;
+      width: 100%;
+    }
+  }
+
+  .jobtitle {
+    flex-grow: 1;
+
+    p {
+      overflow: hidden;
+      text-overflow: ellipsis;
+      display: -webkit-box;
+      -webkit-line-clamp: 2; /* number of lines to show */
+      line-clamp: 2;
+      -webkit-box-orient: vertical;
+    }
+  }
+
+  .description {
+    align-items: baseline;
+    display: flex;
+    flex-direction: row;
+    gap: 0.2rem;
+  }
+
+  h3,
+  p {
+    color: var(--miles_primary_light);
+    margin: 0;
+  }
+}
+
+@keyframes fadeIn {
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+}
